shelterInfoDisplayUsingJS
I wrote my changes on there,

but mainly JS was used to lessen the load on the shelters.html. I duplicated a new page called shelterv2.html.
I haven't checked if add a listing works but its on the shelterv2 page. One last thing, is that the fonts of the listed shelters are not fixed, you can check the style page.